test - change test class name, StockDividendServiceUnitTest to StockDividendServiceTest
